fix(plugin): resolve all compiler warnings with juce_recommended_warning_flags
Fix 190+ compiler warnings to achieve zero-warning builds with strict
JUCE warning flags enabled. Changes include:

- Add explicit static_cast<float>() for int-to-float conversions in
  graphics code (bounds, coordinates, progress bars)
- Add static_cast<size_t>() and static_cast<unsigned int>() for
  sign-conversion warnings in container operations
- Fix variable shadowing by renaming conflicting local variables
- Add missing enum cases to switch statements (FilterType::All)
- Rename shadowed lambda captures and constructor parameters

Warning categories fixed:
- -Wimplicit-int-float-conversion (100+ instances)
- -Wsign-conversion (60+ instances)
- -Wshadow-uncaptured-local (17 instances)
- -Wswitch-enum (5 instances)
- -Wshadow (5 instances)

+++ plugin/src/audio/KeyDetector.cpp
@@ -147,7 +147,7 @@ public:
 
       // Calculate number of samples
       int numSamples = buffer.getNumSamples();
-      audioData.addToSampleCount(numSamples);
+      audioData.addToSampleCount(static_cast<unsigned int>(numSamples));
 
       // Mix to mono and copy to AudioData
       for (int i = 0; i < numSamples; ++i) {
@@ -158,7 +158,7 @@ public:
         sample /= static_cast<float>(numChannels);
 
         // libkeyfinder expects samples in range [-1, 1]
-        audioData.setSample(i, static_cast<double>(sample));
+        audioData.setSample(static_cast<unsigned int>(i), static_cast<double>(sample));
       }
 
       // Run key detection
